refactor: 重构 Label 相关模块，将饼图 label 的布局调整到 LabelLayout 中
