+++ src/geometry/label/index.ts
@@ -1,15 +1,17 @@
 import { BBox, IGroup, IShape } from '../../dependents';
 import { LooseObject } from '../../interface';
 import { GeometryLabelConstructor } from './base';
+import { LabelItem } from './interface';
 
 /**
  * label 布局函数定义
+ * @param items 存储每个 label 的详细信息
  * @param labels 所有的 labels 图形实例
  * @param shapes 所有 label 对应的图形元素实例
  * @param region 画布区域
  * @param cfg 用于存储各个布局函数开放给用户的配置数据
  */
-type GeometryLabelsLayoutFn = (labels: IGroup[], shapes: IShape[] | IGroup[], region: BBox, cfg?: LooseObject) => void;
+type GeometryLabelsLayoutFn = (items: LabelItem[], labels: IGroup[], shapes: IShape[] | IGroup[], region: BBox, cfg?: LooseObject) => void;
 
 const GEOMETRY_LABELS_MAP: Record<string, GeometryLabelConstructor> = {};
 const GEOMETRY_LABELS_LAYOUT_MAP: Record<string, GeometryLabelsLayoutFn>  = {};
